requirements.txt
flask
flask_jwt_extended
python_dotenv
pymongo